% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapters.R
\name{as_mvpa_dataset}
\alias{as_mvpa_dataset}
\alias{as_mvpa_dataset.fmri_dataset}
\title{Convert to rMVPA dataset}
\usage{
as_mvpa_dataset(x, ...)

\method{as_mvpa_dataset}{fmri_dataset}(x, mask = NULL, ...)
}
\arguments{
\item{x}{object to convert (e.g., fmri_dataset)}

\item{...}{additional arguments passed to methods}

\item{mask}{optional mask to use (defaults to x$mask)}
}
\value{
an mvpa_dataset object
}
\description{
Generic function to convert various fMRI data objects to rMVPA's mvpa_dataset format
}
\section{Methods (by class)}{
\itemize{
\item \code{as_mvpa_dataset(fmri_dataset)}: Convert fmri_dataset to mvpa_dataset

}}
