% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{compute_metrics}
\alias{compute_metrics}
\title{Compute a set of metrics from probs/y_true}
\usage{
compute_metrics(probs, y_true, metrics = c("accuracy", "auc"))
}
\description{
Compute a set of metrics from probs/y_true
}
\keyword{internal}
