% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{.get_run_ids_from_test_data}
\alias{.get_run_ids_from_test_data}
\title{Get run IDs for test data (internal helper)}
\usage{
.get_run_ids_from_test_data(fit_obj, Y_test, ev_model_test)
}
\description{
Get run IDs for test data (internal helper)
}
\keyword{internal}
