% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{estimate_hrf_theta}
\alias{estimate_hrf_theta}
\title{Estimate HRF basis coefficients from soft labels}
\usage{
estimate_hrf_theta(X_list, Z_event, hrf_obj, penalty = 0.01)
}
\description{
Estimate HRF basis coefficients from soft labels
}
\keyword{internal}
