% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{hrf_metrics}
\alias{hrf_metrics}
\title{Compute metrics from prediction probabilities and true labels}
\usage{
hrf_metrics(probs, y_true, metrics = c("accuracy", "auc"))
}
\arguments{
\item{probs}{matrix of probabilities (N x K)}

\item{y_true}{factor of true labels with same levels as colnames(probs)}

\item{metrics}{character vector of metrics: "accuracy", "auc", "logloss", "brier", "bal_acc"}
}
\description{
Compute metrics from prediction probabilities and true labels
}
