% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{prepare_decoder_inputs}
\alias{prepare_decoder_inputs}
\title{Prepare decoder design inputs}
\usage{
prepare_decoder_inputs(ev_model, hrf = NULL, background = TRUE)
}
\description{
Prepare decoder design inputs
}
\keyword{internal}
